elevator PID in auton

+++ src/com/team2502/robot2018/Constants.java
@@ -13,12 +13,12 @@ public class Constants
     /**
      * How high the elevator must be in order to put a cube in the switch
      */
-    public static final double SWITCH_ELEV_HEIGHT_FT = 1.5F;
+    public static final float SWITCH_ELEV_HEIGHT_FT = 1.5F;
 
     /**
      * How high the elevator must be in order to put a cube in the scale
      */
-    public static final double SCALE_ELEV_HEIGHT_FT = 7F;
+    public static final float SCALE_ELEV_HEIGHT_FT = 7F;
